LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/send-empty-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/echo' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 57 - Socket is not connected)
WebSocket: Calling send() without arguments should raise SyntaxError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.send() threw exception TypeError: Not enough arguments.
FAIL closeEvent.wasClean should be true. Was false.
PASS receivedMessages.length is 1
PASS receivedMessages[0] is "Goodbye"
PASS successfullyParsed is true

TEST COMPLETE
